python-fast-api-htmx-sqlite/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiosqlite==0.20.0
    # via -r requirements.in
alembic==1.13.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio[trio]==3.7.1
    # via
    #   fastapi-login
    #   httpx
    #   starlette
async-generator==1.10
    # via trio
attrs==24.2.0
    # via
    #   outcome
    #   trio
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
cryptography==43.0.0
    # via python-jose
dnspython==2.6.1
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via pydantic
fastapi==0.112.0
    # via
    #   -r requirements.in
    #   fastapi-login
    #   fastapi-sso
fastapi-login==1.10.1
    # via -r requirements.in
fastapi-sso==0.15.0
    # via -r requirements.in
greenlet==3.0.3
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via fastapi-sso
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   trio
itsdangerous==2.1.2
    # via -r requirements.in
jinja2==3.1.4
    # via -r requirements.in
mako==1.3.5
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
    #   werkzeug
nulltype==2.3.1
    # via plaid-python
oauthlib==3.2.2
    # via fastapi-sso
outcome==1.3.0.post0
    # via trio
plaid-python==24.0.0
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-sso
pydantic-core==2.20.1
    # via pydantic
pyjwt==2.9.0
    # via fastapi-login
pytailwindcss==0.2.0
    # via -r requirements.in
pytelegrambotapi==4.23.0
    # via telebot
python-dateutil==2.9.0.post0
    # via plaid-python
python-dotenv==1.0.1
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via -r requirements.in
requests==2.32.3
    # via
    #   pytelegrambotapi
    #   telebot
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   trio
sortedcontainers==2.4.0
    # via trio
sqlalchemy==2.0.32
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via
    #   -r requirements.in
    #   fastapi
telebot==0.0.5
    # via -r requirements.in
trio==0.21.0
    # via anyio
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   alembic
    #   fastapi
    #   fastapi-login
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.2
    # via
    #   plaid-python
    #   requests
uvicorn==0.30.6
    # via -r requirements.in
werkzeug==3.0.1
    # via -r requirements.in
